// client/src/module_bindings/shape_type_type.ts
// THIS FILE IS AUTOMATICALLY GENERATED BY SPACETIMEDB. EDITS TO THIS FILE
// WILL NOT BE SAVED. MODIFY TABLES IN YOUR MODULE SOURCE CODE INSTEAD.

/* eslint-disable */
/* tslint:disable */
// @ts-nocheck
import {
  AlgebraicType,
  AlgebraicValue,
  BinaryReader,
  BinaryWriter,
  ConnectionId,
  DbConnectionBuilder,
  DbConnectionImpl,
  Identity,
  ProductType,
  ProductTypeElement,
  SubscriptionBuilderImpl,
  SumType,
  SumTypeVariant,
  TableCache,
  TimeDuration,
  Timestamp,
  deepEqual,
} from "@clockworklabs/spacetimedb-sdk";

import type {
  CallReducerFlags,
  DbContext,
  ErrorContextInterface,
  Event,
  EventContextInterface,
  ReducerEventContextInterface,
  SubscriptionEventContextInterface,
} from "@clockworklabs/spacetimedb-sdk";
// A namespace for generated variants and helper functions.
export namespace ShapeType {
  // These are the generated variant types for each variant of the tagged union.
  // One type is generated per variant and will be used in the `value` field of
  // the tagged union.
  export type Circle = { tag: "Circle" };
  export type Rectangle = { tag: "Rectangle" };
  export type Line = { tag: "Line" };
  export type Polygon = { tag: "Polygon" };
  export type Text = { tag: "Text" };

  // Helper functions for constructing each variant of the tagged union.
  // ```
  // const foo = Foo.A(42);
  // assert!(foo.tag === "A");
  // assert!(foo.value === 42);
  // ```
  export const Circle = { tag: "Circle" };
  export const Rectangle = { tag: "Rectangle" };
  export const Line = { tag: "Line" };
  export const Polygon = { tag: "Polygon" };
  export const Text = { tag: "Text" };

  export function getTypeScriptAlgebraicType(): AlgebraicType {
    return AlgebraicType.createSumType([
      new SumTypeVariant("Circle", AlgebraicType.createProductType([])),
      new SumTypeVariant("Rectangle", AlgebraicType.createProductType([])),
      new SumTypeVariant("Line", AlgebraicType.createProductType([])),
      new SumTypeVariant("Polygon", AlgebraicType.createProductType([])),
      new SumTypeVariant("Text", AlgebraicType.createProductType([])),
    ]);
  }

  export function serialize(writer: BinaryWriter, value: ShapeType): void {
      ShapeType.getTypeScriptAlgebraicType().serialize(writer, value);
  }

  export function deserialize(reader: BinaryReader): ShapeType {
      return ShapeType.getTypeScriptAlgebraicType().deserialize(reader);
  }

}

// The tagged union or sum type for the algebraic type `ShapeType`.
export type ShapeType = ShapeType.Circle | ShapeType.Rectangle | ShapeType.Line | ShapeType.Polygon | ShapeType.Text;

export default ShapeType;